Added basic username/pass and github auth
It's still in a rough shape, needs more tests and
beautification of haml templates.

There is a current_user method that returns the User
object of the current user or nil if the user is not
authenticated.

A User object has the admin? and new_user? methods.
The first user that gets created has the admin role.
Any following users created have the new_user role.